simplified encryption and peerlist
