+++ internal/p2p/peerList_handler.go
@@ -12,47 +12,52 @@ import (
 	"github.com/multiformats/go-multiaddr"
 )
 
-// Returns a list of `peerID, peerAddr\n` for every peer in the peer list in order
+// Returns a list of `peerID peerAddr\n` for every peer in the peer list in order
 func (p *GokerPeer) getPeerList() string {
 	p.peerListMutex.Lock()
 	defer p.peerListMutex.Unlock()
 
-	var peerList string
+	var peerList []string
 	for _, peerInfo := range p.peerList {
-		peerList += fmt.Sprintf("%s %s\n", peerInfo.ID.String(), peerInfo.Addr.String()) // Sends peer list as a multi-line string
+		peerList = append(peerList, fmt.Sprintf("%s %s", peerInfo.ID.String(), peerInfo.Addr.String()))
 	}
-	return peerList
+	return strings.Join(peerList, "\n")
 }
 
 // Set the peer list (Given the output from the getPeerList function) and connect to all new peers
 func (p *GokerPeer) setPeerListAndConnect(peerList string) {
-	p.peerListMutex.Lock()
-	defer p.peerListMutex.Unlock()
 
 	var sentPeerList []peerInfo
+	scanner := bufio.NewScanner(strings.NewReader(peerList))
 
-	scanner := bufio.NewScanner(strings.NewReader(peerList)) // Put it in a scanner for processing
-	for scanner.Scan() {                                     // Next token
-		line := scanner.Text()            // Get the current line seperated by \n
-		parts := strings.Split(line, " ") // Split them to peer ID - Multiaddr
+	for scanner.Scan() {
+		parts := strings.Fields(scanner.Text()) // Handles spaces in the line
 		if len(parts) < 2 {
 			continue
 		}
 
 		newPeerID, err := peer.Decode(parts[0]) // make it a peerID
 		if err != nil {
-			log.Printf("setPeerList: Unable to decode incoming peerID")
+			log.Fatal("setPeerList: Unable to decode incoming peerID")
 		}
 
 		addr, err := multiaddr.NewMultiaddr(parts[1]) // Make it a multiaddr
 		if err != nil {
-			log.Println("setPeerList: Unable to create multiaddr for incoming peerID - " + newPeerID.String())
+			log.Fatal("setPeerList: Unable to create multiaddr for incoming peerID - " + newPeerID.String())
 		}
 
 		sentPeerList = append(sentPeerList, peerInfo{ID: newPeerID, Addr: addr})
 	}
 
-		
+	if err := scanner.Err(); err != nil {
+		log.Fatalf("setPeerList: Error reading peer list - %v", err)
+	}
+
+	// Time to set it
+
+	p.peerListMutex.Lock()
+	defer p.peerListMutex.Unlock()
+
 	existing := make(map[peerInfo]struct{}, len(p.peerList))
 	for _, peerInfo := range p.peerList {
 		existing[peerInfo] = struct{}{}
@@ -62,7 +67,8 @@ func (p *GokerPeer) setPeerListAndConnect(peerList string) {
 	for _, newPeerInfo := range sentPeerList {
 		if _, found := existing[newPeerInfo]; !found {
 			p.peerList = append(p.peerList, newPeerInfo)
-			
+			existing[newPeerInfo] = struct{}{} // Prevent duplicate additions
+
 			// Connect to each peer new peer
 			if newPeerInfo.ID != p.ThisHost.ID() {
 				addrInfo := peer.AddrInfo{ID: newPeerInfo.ID, Addrs: []multiaddr.Multiaddr{newPeerInfo.Addr}}
@@ -73,10 +79,5 @@ func (p *GokerPeer) setPeerListAndConnect(peerList string) {
 				}
 			}
 		}
-	}
-
-
-	if err := scanner.Err(); err != nil {
-		log.Printf("setPeerList: Error reading peer list - %v", err)
 	}
 }
